Move `addTransceiversIfNeeded` to a separate file
